HW4-2.5-11

+++ ExcerciseTest.py
@@ -3,7 +3,7 @@ from scipy.misc import derivative
 
 x = float(1.5)
 TOL = float(10 ** (-4))
-N1 = float(100)
+N1 = float(20)
 f = lambda x: x ** float(2) - float(2)
 f1 = lambda x: float(2) * x
